Third Android Exercise

+++ app/src/main/java/at/fh/swengb/loggingviewsandactivity/LessonRating.kt
@@ -1,6 +1,6 @@
 package at.fh.swengb.loggingviewsandactivity
 
-class LessonRating(var ratingValue: Int, var feedback: String) {
+class LessonRating(var ratingValue: Double, var feedback: String) {
 
 
 }
